feat(app): Add a method to remove password
Remove password method
Solve error if file doesn't exists

+++ packages/core/src/readPasswds.ts
@@ -8,17 +8,19 @@ export default (password: string) => {
     decFiles(path.join(os.homedir(), '.hider.datas'), password);
     const datas: { whereToUse: string; password: string }[] = [];
     setTimeout(() => {
-      const file = fs.readFileSync(
-        path.join(os.homedir(), '.hider.tmp'),
-        'utf8'
-      );
-      fs.unlinkSync(path.join(os.homedir(), '.hider.tmp'));
-      const lines = file.split('\n');
-      for (let i = 0; i < lines.length - 1; i++) {
-        datas.push({
-          whereToUse: lines[i].split(':')[0],
-          password: lines[i].split(':')[1],
-        });
+      if (fs.existsSync(path.join(os.homedir(), '.hider.tmp'))) {
+        const file = fs.readFileSync(
+          path.join(os.homedir(), '.hider.tmp'),
+          'utf8'
+        );
+        fs.unlinkSync(path.join(os.homedir(), '.hider.tmp'));
+        const lines = file.split('\n');
+        for (let i = 0; i < lines.length - 1; i++) {
+          datas.push({
+            whereToUse: lines[i].split(':')[0],
+            password: lines[i].split(':')[1],
+          });
+        }
       }
       resolve(datas);
     }, 100);
